Rename '/api/schedule-visit' to '/api/book-a-visit'
